Refactored litmusctl

+++ pkg/types/api_types.go
@@ -3,7 +3,7 @@ package types
 
 type AuthResponse struct {
 	AccessToken string `json:"access_token"`
-	ExpiresIn   int    `json:"expires_in"`
+	ExpiresIn   int64    `json:"expires_in"`
 	Type        string `json:"type"`
 }
